Finished generating the cube properly.  Seems to work well, if a bit messily. Added a method for counting the number of rendered cubes, since I thought I was dduplicating cubes during generation (and I was!). Started writing a face rotation function, but..:
This is great and all, but I think I'm going about this wrong, as we're not solving the cube by analysing it visually and dedtermining what to do from there - we're solving a matrix-representation of the cube and updating the modell to reflect this.
Consequently, the next thing I'm going to try is implement a Rubik's Cube matrix and update the Rubiks Cube class's primitives to reflect the state of the matrix.