Finished the cross algorithm for the first face. This currently involves repeating the algorithm until we don't have to change anything. Realised during this that I'd not done `sideNeighboursTemp` correctly - hey ho.
Fixed the shuffle trying to rotate a side with depth beyond its half-way point.
Implemented move recording (saved in `lastMoves.txt`), as well as the facility to replay those moves.
The idea here is that you can copy the moves you want out of the file (such as the shuffling moves), then replay those onto the cube (such as to re-create an issue).

Ghosted some more functions, that we'll probably be writing next!
Some more cleanup stuff.